+++ src/RestApi/BulkOperation/BulkOperation.types.ts
@@ -1,6 +1,7 @@
 import { InputSource } from '../Device/InputSource';
-import { TResolutionItem } from '../Device/Resolution/IDeviceResolution';
-import { ApplicationType, DeviceActionType, Orientation, VideoOrientation } from './enums';
+import { ApplicationType, DeviceActionType, Orientation, VideoOrientation } from './BulkOperation.enums';
+import { ResolutionItem } from '../Device/IDevice';
+import { TelemetryCheckIntervals } from '../Device/Monitoring/DeviceMonitoring.types';
 
 export interface IConfigValues {
 	platformUri?: string | null;
@@ -60,7 +61,8 @@ export type LogData = {
 	[DeviceActionType.UNINSTALL_PACKAGE]: {
 		packageName: string;
 		applicationType: ApplicationType;
-		specs: Record<string, unknown>;
+		// eslint-disable-next-line @typescript-eslint/ban-types
+		specs: object;
 	};
 	[DeviceActionType.POWER_ACTION]: {
 		powerType: string;
@@ -77,15 +79,9 @@ export type LogData = {
 		enabled: boolean;
 	};
 	[DeviceActionType.RESIZE]: {
-		resolution: TResolutionItem;
+		resolution: ResolutionItem;
 		orientation: Orientation;
 		videoOrientation?: VideoOrientation;
-	};
-	[DeviceActionType.PROVISION]: {
-		verificationHash: string;
-	};
-	[DeviceActionType.DEPROVISION]: {
-		verificationHash: string;
 	};
 	[DeviceActionType.UPDATE_NAME]: {
 		name: string;
@@ -98,37 +94,28 @@ export type LogData = {
 	[DeviceActionType.CREATE_TIMING]: {
 		appletUid: string;
 		appletVersion: string;
-		startsAt: Date;
-		endsAt: Date;
 		configuration: Record<string, unknown>;
-		finishEvent: {
-			type: 'DURATION' | 'IDLE_TIMEOUT' | 'SCREEN_TAP';
-			data?: any;
-		};
-		position: number;
 	};
 	[DeviceActionType.UPDATE_TIMING]: {
 		appletUid: string;
 		appletVersion: string | undefined;
-		startsAt: Date | undefined;
-		endsAt: Date | undefined;
 		configuration: Record<string, unknown> | undefined;
-		finishEvent: {
-			type: 'DURATION' | 'IDLE_TIMEOUT' | 'SCREEN_TAP' | undefined;
-			data?: any | undefined;
-		} | undefined;
-		position: number | undefined;
+		configurationSet: Record<string, unknown> | undefined;
+		configurationRemoveKeys: string[] | undefined;
 	};
 	[DeviceActionType.DELETE_TIMING]: {
+		uid: string;
 		appletUid: string;
-		appletVersion: string | undefined;
+		appletVersion: string;
 	};
 	[DeviceActionType.SET_DEVICE_APPLET_TEST_SUITE]: {
 		appletUid: string;
 		appletVersion: string;
 		tests: string[];
 	};
-	[DeviceActionType.SET_TEST_SUITE]: Record<string, never>;
+	[DeviceActionType.SET_TEST_SUITE]: {
+		tests: string[];
+	};
 	[DeviceActionType.START_PACKAGE]: {
 		packageName: string;
 		applicationType: ApplicationType;
@@ -179,6 +166,37 @@ export type LogData = {
 	};
 	[DeviceActionType.SET_ORGANIZATION]: {
 		organizationUid: string;
+	};
+	[DeviceActionType.SET_AUTO_RECOVERY]:
+		| {
+				enabled: true;
+				/** It defines in miliseconds period whose elapsing leads to mark browser process as unresponding. */
+				healthcheckIntervalMs: number;
+	}
+		| {
+				enabled: false;
+				/** When specified it defines period until automatic enabling auto recovery process in miliseconds. */
+				autoEnableTimeoutMs?: number;
+	};
+	[DeviceActionType.SET_PEER_RECOVERY]:
+		| {
+				enabled: true;
+	}
+		| {
+				enabled: false;
+				/** When specified it defines period until automatic enabling peer recovery process in miliseconds. */
+				autoEnableTimeoutMs?: number;
+	};
+	[DeviceActionType.ENABLE_EXTENDED_TELEMETRY]: {
+		deviceIdentityHash: string;
+		duration: number;
+	};
+	[DeviceActionType.DISABLE_EXTENDED_TELEMETRY]: {
+		deviceIdentityHash: string;
+	};
+	[DeviceActionType.TELEMETRY_INTERVALS]: {
+		deviceIdentityHash: string;
+		telemetryCheckIntervals: TelemetryCheckIntervals;
 	};
 };