+++ src/RestApi/Timing/Timing.types.ts
@@ -1,0 +1,5 @@
+export enum FinishEventType {
+	DURATION = 'DURATION',
+	IDLE_TIMEOUT = 'IDLE_TIMEOUT',
+	SCREEN_TAP = 'SCREEN_TAP',
+}