+++ CHANGELOG.md
@@ -3,9 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
-
 ## [Unreleased]
+### Added
+- Finalize bulk operations implementation
+
 ### Fixed
+- Fix timings comparator in `update` timing function
 - `setTelemetryIntervals` returns error when value is more than 1 year
 
 ## [1.19.2] - 2023-04-04